Add data pipeline prototype
The data-pipeline downloads, filters, normalizes, and serializes both
station and line content from VBB modules to create a stable source for
the game.

The idea is that the pipeline would run locally only once, and remotely
on every build, generating content to the "dist/data/" folder with
the "stations.json" and "lines.json" files. The web-app will then
consume the files to feed internal needs.

That's all possible thanks to @derhuerst work on the https://github.com/derhuerst/vbb-modules

💛 Thanks a lot!