Extract View and State to subpackages
In order to better manage the state and the view, I'm trying to extract
to indivual modules.

However, it is noticeable that almost all game related files leak data
or structure to other files...

Also add a lot of tests, covering almost all scenarios for the State
itself and the update function.