Implement timeout game mode
Now, rather than a fixed number of questions, we set a time interval for
the players to really show their skills under pressure.

In order to do that some re-design was necessary, for that reason you
might see a lot of CSS/HTML related changes here.

Important enough is that the core code changes required for the new
feature are quite small, 5-10 lines, injecting a new update message with
the clock tick and manipulating the game state when there is no time
left.